Move to static
